System call lseek() tested
lseek almost fine, only 1 point works bad (neg value on truncate case

+++ kern/include/syscall.h
@@ -71,7 +71,7 @@ int sys_open(userptr_t path, int openflags, mode_t mode, int *errp);
 int sys_close(int fd, int *errp);
 int sys_write(int fd, userptr_t buf_ptr, size_t size, int *errp);
 int sys_read(int fd, userptr_t buf_ptr, size_t size,  int *errp);
-int sys_lseek(int fd, off_t offset, int whence, int *errp);
+off_t sys_lseek(int fd, off_t offset, int whence, int *errp);
 int sys_chdir(const char *pathname, int *errp);
 int sys_dup2(int oldfd, int newfd, int *errp);
 void sys__exit(int status);
